backend/requirements.txt
fastapi==0.95.0
uvicorn==0.18.3
pymongo==4.1.1
apscheduler==3.8.1
llama-cpp-python==0.1.5
langchain==0.0.133
langchain_huggingface==0.0.3
feedparser==6.0.10
sentence-transformers>=2.6.0
langchain==0.0.133
faiss-cpu
beautifulsoup4
lxml
python-dateutil
